+++ subgraphs/mvp/src/mappings/phuture/vToken.ts
@@ -1,8 +1,9 @@
 import { VTokenTransfer, UpdateDeposit } from '../../types/templates/vToken/vToken';
-import { convertTokenToDecimal, loadOrCreateIndexAsset, loadOrCreateVToken } from '../entities';
+import { loadOrCreateIndexAsset, loadOrCreateVToken } from '../entities';
 import { updateDailyAssetStat, updateStat } from './stats';
 import { Asset, Index, vToken } from '../../types/schema';
 import { Address, BigDecimal, BigInt, log } from '@graphprotocol/graph-ts';
+import { convertTokenToDecimal } from '../../utils/calc';
 
 export function handlerVTokenTransfer(event: VTokenTransfer): void {
   let vt = vToken.load(event.address.toHexString());
@@ -75,9 +76,9 @@ export function handlerUpdateDeposit(event: UpdateDeposit): void {
   let vt = loadOrCreateVToken(event.address);
 
   if (event.params.depositedAmount.gt(BigInt.zero())) {
-    vt.assetReserve = vt.assetReserve.plus(event.params.depositedAmount.toBigDecimal());
+    vt.assetReserve = vt.assetReserve.plus(event.params.depositedAmount);
   } else {
-    vt.assetReserve = vt.assetReserve.minus(vt.deposited.toBigDecimal());
+    vt.assetReserve = vt.assetReserve.minus(vt.deposited);
   }
 
   vt.deposited = event.params.depositedAmount;
